+++ src/middleware/authnJwt.js
@@ -1,17 +1,25 @@
+const { ApiError } = require('../api/util')
 const safeAsync = require('../middleware/safeAsync')
 const { getUserFromJwt } = require('../auth/util')
 
 module.exports = safeAsync(async (req, res, next) => {
+  if (res.locals.userAuthn) {
+    // Something else in the chain already handled authn
+    next()
+    return
+  }
+
   const jwtCookie = req.cookies.jwt
   if (!jwtCookie) {
-    res.status(401).send()
+    // Maybe they authenticated a different way
+    next()
     return
   }
 
   const user = await getUserFromJwt(jwtCookie)
   if (user === null) {
     // This shouldn't ever happen, but if it does...
-    res.status(401).send()
+    next(new ApiError(401, 'Invalid cookie'))
     return
   }